fix: Remove the duplication of addEventListener (#232)
The lines from 172 to 191 seem to be exactly the same with the lines from 206 to 227. As they are in the same scope, I guess the same handler is being assigned twice. 
**So I erased the lines 172-191.**
If it was intended, sorry for bothering. 
Btw it's an amazing example, I'm learning a lot from this one. Thanks for the great job!