Allow signalling server to restart without interrupting audio/video.
A grpc stream can now identify itself with a string id. Identified
streams do not close their peer connections when the grpc connection is
interrupted. A grpc stream that identifies as an existing peer will
reconnect to that peer. Peers are deleted when they go to ICE state
closed.

This allows the signalling server to restart and reconnect to ion-sfu
without interrupting user audio/video.

A grpc stream that does not identify behaves exactly as before.