Set `DownTrack` enabled status on create
Previously, new `DownTrack`s were enabled when their `Bind` was called.
However it is possible that a peer has already attempted to mute/unmute
the track before this point, which resulted in that status being lost.
This would be the case if a client tried to mute/unmute a track from an
`ontrack` event, for example.

This changes it to set the initial status when the `DownTrack` is
created instead, so that any subsequent changes will be honored.